cmake: add support for sanitizers
This adds ubsan/asan to the Debug build
when they are available on the system.

While asan is fast for a memory checker,
there is still a 2-3x slowdown compared
to running without it. Always adding asan
when it is available keeps the build system
simple though, the release build should be used
when performance is a concern.