First runnable sketch of event library.
The library uses a std::priority_queue to hold EventStreams, each of
which generate a sequence of events. The baseclass of events is Event
and each application should subclass it to add application specific
data. Similarly, EventStream is the baseclass of event streams and is
sub classed to add the specific event generation functionality.