Add Kafka support
This adds a basic event that sends Kafka
messages when triggered.

The event trigger sets up the Kafka connection
and uses a hardcoded configuration to so the
input side has something to do with its input.
This commit does not add anything that uses the
object. These limitations will be lifted in later
commits.