Provide opaque constructors for Input subclassses
This allows us to stop exposing the internal details
of the Input subclasses and Event/EventStream subclasses.
This makes main.cpp into a straight consumer of
the library without any implementation parts.
The changes also allows moving the static xmlInput variable
into XMLInput.cpp and manage the lifetime of that object
by the XMLFileInput object instead of the TurboEvents
object.

There are still internal details exposed in the interface
so as an intermediate step this commit removes the
private declaration of TurboEvents::greaterES and
the priority queue. These are still considered private
variables of the TurboEvents class and should not
be used from main. They will be moved into the private
parts of the library in a future commit.