Put the countdown stream in separate files
A countdown stream has no real world use,
but this code is our current example of algorithmic
generation of events so put it in the right
place in the tree and keep it around for
ensuring we don't get regressions in functionality.

Also divide the used interval values by 5.
This saves us 8 seconds in the CI runs.