Add functionality to time shift file input.
The timestamps of events in files are absolute. This requires
the files to be updated for each use, which is inconvenient.
This patch introduces the command line flag --timeshift which
tells turboevents to shift the time stamps in each file by a
constant amount so that the first event in the file is triggered
approximately when the program starts.

The test files are also updated to give a reasonable execution
time for testing with the new flag.