Base event time on previous time in CountDownEventStream.
This patch ensures that event times are evenly spaced for
CountDownEventStream by basing each time stamp on the
previous time stamp rather than clock time when the event
is generated. This ensures expected behaviour when time
scaling is implemented in a future commit.

The constructor for EventStream is also simplified to
always initialize next to nullptr.