Add a general input container
This adds an input container that can
be used from the Python API. This turned
out to be the same container that XMLInput
has already implemented so move the relevant
XMLInput code into a separate file and use
that from XMLInput.

The natural API would be

createContainerInput(std::vector<std::unique_ptr<>> evs)

but that does not work with Pybind, so remove
the parameter and hide a container inside
TurboEventsImpl that can be accessed through
addEvent() in the public API.

This commit also groups the related methods
and sorts them in alphabetical order.