Add command line option to scale time.
This patch adds the option '--scale <x>' which scales the
intervals between generated events by a factor <x>. A scaling
factor of 1 makes no change, a factor above 1 increases the
(real) time between event delivery while a factor below 1
speeds up delivery. A factor of 0 is allowed and delivers
the events as fast as possible.