Add support for multiple outputs
This converts the single output into
a container of multiple outputs that
events are sent to. The key insight
is that we need to separate the event data
and the actor that uses the data.

The consequence of that movement is that we
get mulitple output objects, so we also get
multiple starttime/timeshift statuses. Move
that information into a new class "Config"
that we only have a single instance of
(the TurboEventsImpl object). This class
is also a conveient place for the "makeEvent"
methods.

There is nothing that prevents having the same output
multiple times, so:

turboevents test/*xml -output='print,print'

will print the data to stdout twice and:

turboevents test/*xml -output='print,kafka,print'

will print the data, send it through kafka, and
then print it again.

Beyond the functional improvement, the net result
of the change is removing a bunch of overloading
and 5% of the lines of code.